Move os_event definition to separate header file
Summary:
This is in preparation to porting the commit to 8.0.
https://github.com/facebook/mysql-5.6/commit/c1b8d839a740618656dde6a28ec8a6a914b200fc
The definition is needed because the struct os_event will be inlined
in rwlock and mutex structures in subsequent commit.

Differential Revision: D25364668
